requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pyOpenSSL==23.0.0
Werkzeug==2.2.2
django-extensions==3.2.1
PyLTI1p3==2.0.0
django-cors-headers==3.13.0
psycopg[binary]
django[argon2]
imagekitio
StrEnum
pycryptodome
